requirements.txt
numpy
matplotlib
tensorflow      # will pull CPU version by default
scikit-learn
pillow
streamlit
streamlit-drawable-canvas
